Generating IETF YANG and openconfig extensions
