--- napalm_yang/openconfig_extensions/oc_ext.py
@@ -1,9 +1,0 @@
-"""
-This module provides extensions to the YANG language to allow
-OpenConfig specific functionality and meta-data to be defined.
-"""
-
-
-class OpenConfigExtensions():
-    def _init__(self):
-        self.version = None
